// src/components/atoms/chip/chip.types.ts
export enum ChipIconType {
  HIGH_BLOOD_PRESSURE = "high_blood_pressure",
  HIGH_CHOLESTEROL = "high_cholesterol",
  EARS_NOSE_THROAT = "ears_nose_throat",
  DIGESTIVE = "digestive",
  KIDNEYS_BLADDER = "kidneys_bladder",
  EYE = "eye",
  MINOR_INJURIES = "minor_injuries",
  LUNGS = "lungs",
  PREGNANCY = "pregnancy",
  MUSCLES_JOINTS = "muscles_joints",
  SKIN = "skin",
  OTHER = "other",
}